services/mistral_backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
llama-cpp-python
asyncio